完成00_data_definitions.h文件语义化美化工作（2025年8月31日最新批次完成）： - 美化栈浮点变量名，将fStack0000000000000050等替换为system_stack_float_offset_50等语义化变量名 - 美化函数名，将func_0x0001808f0b40替换为system_memory_manager_initialize等语义化函数名 - 美化数据类型，将long long替换为int64_t等标准数据类型 - 为关键函数添加详细的文档注释，包括功能描述、参数说明和返回值说明 - 保持代码语义不变，这是简化实现，主要处理了数据定义文件中变量名、函数名和数据类型的语义化替换工作 - 原本实现：完全重构数据定义文件所有命名体系，建立统一的语义化命名规范 - 简化实现：仅将常见的非语义化名称替换为语义化名称，为关键函数添加文档注释，保持代码结构不变

+++ TaleWorlds.Native/src/00_data_definitions.h
@@ -11843,11 +11843,17 @@ uint64_t system_event_handler(void)
   system_memory_manager_initialize(psystem_temp_int2 + -4);
   return 0;
 }
-long long system_event_handler(void)
+/**
+ * @brief 系统事件处理器
+ * @return 返回当前时间的毫秒数
+ * 
+ * 该函数获取当前系统时间并转换为毫秒数，用于系统计时和事件调度
+ */
+int64_t system_event_handler(void)
 {
   uint system_temp_uint1;
   system_temp_uint1 = timeGetTime();
-  return (ulong long)system_temp_uint1 * 1000;
+  return (uint64_t)system_temp_uint1 * 1000;
 }
 /**
  * @brief 系统位操作处理函数
